pyproject.toml
[project]
name = "github-api-stats-puller"
version = "0.1.0"
description = "GitHub API statistics puller and metrics service"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pygithub>=2.1.0",
    "pydantic>=2.5.0",
    "python-dateutil>=2.8.2",
    "plotly>=5.17.0",
    "loguru>=0.7.0",
    "python-dotenv>=1.1.1",
]

[project.scripts]
github-stats = "github_stats.main:main"
github-api = "github_stats.app:main"

[tool.uv]
package = true
